fix: update tests and fix PR 6334 merge
