+++ ddtrace/appsec/handlers.py
@@ -91,6 +91,7 @@ def _on_request_span_modifier(request, environ, _HAS_JSON_MIXIN, exception_type)
 
 
 def _on_request_init(wrapped, instance, args, kwargs):
+    wrapped(*args, **kwargs)
     if _is_iast_enabled():
         try:
             from ddtrace.appsec.iast._taint_tracking import OriginType
